Fix panel layout UI, add color config for labels

+++ src/main/java/malte0811/controlengineering/controlpanels/PanelComponentInstance.java
@@ -3,7 +3,7 @@ package malte0811.controlengineering.controlpanels;
 import com.mojang.datafixers.util.Pair;
 import com.mojang.serialization.Codec;
 import malte0811.controlengineering.bus.BusState;
-import malte0811.controlengineering.util.math.Vec2i;
+import malte0811.controlengineering.util.math.Vec2d;
 import malte0811.controlengineering.util.typereg.TypedInstance;
 import net.minecraft.network.PacketBuffer;
 import net.minecraft.util.ActionResultType;
@@ -102,7 +102,7 @@ public final class PanelComponentInstance<Config, State> extends TypedInstance<P
         currentState = Pair.of(newConfig, getState());
     }
 
-    public Vec2i getSize() {
+    public Vec2d getSize() {
         return getType().getSize(getConfig());
     }
