+++ src/main/java/malte0811/controlengineering/controlpanels/components/Indicator.java
@@ -7,7 +7,8 @@ import com.mojang.serialization.Codec;
 import malte0811.controlengineering.ControlEngineering;
 import malte0811.controlengineering.bus.BusState;
 import malte0811.controlengineering.controlpanels.PanelComponentType;
-import malte0811.controlengineering.util.math.Vec2i;
+import malte0811.controlengineering.controlpanels.components.config.ColorAndSignal;
+import malte0811.controlengineering.util.math.Vec2d;
 import net.minecraft.util.ActionResultType;
 import net.minecraftforge.common.Tags;
 
@@ -21,7 +22,7 @@ public class Indicator extends PanelComponentType<ColorAndSignal, Integer> {
         super(
                 ColorAndSignal.DEFAULT, 0,
                 ColorAndSignal.CODEC, Codec.INT,
-                new Vec2i(1, 1),
+                new Vec2d(1, 1),
                 TRANSLATION_KEY
         );
     }
